+++ app/User.php
@@ -2,6 +2,7 @@
 
 namespace App;
 
+use App\Models\Highlight;
 use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
@@ -37,4 +38,11 @@ class User extends Authenticatable implements MustVerifyEmail
     protected $casts = [
         'email_verified_at' => 'datetime',
     ];
+
+    /**
+     * The relationships.
+     */
+    public function highlights () {
+        return $this->hasMany(Highlight::class);
+    }
 }