+++ routes/api.php
@@ -18,6 +18,9 @@ Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
 
+// Route::get('/dashboard', 'Admin\DashboardController@index') -> name('dashboard') ; //-> middleware('can:admin');
+
+
 Route::get('/books','Bible\GetBibleController@getBibleBooks')->name('getBibleBooks');
 Route::get('/chapter/{chapter_id}', 'Bible\GetBibleController@getChapterById');